IMP-124: sorting wallets by creation date (#47)
* IMP-124: natural sort order for wallet_id

* IMP-124: add test

* IMP-124: desc -> asc

* IMP-124: truncate table after test

* IMP-124: fix EmptyLineSeparator

* IMP-124: fix WalletFunctionTest

* IMP-124: fix continuation token pagination for wallets

* IMP-124: fix testGetWallets

* IMP-124: more tests

* IMP-124: sorting wallets by creation date

+++ src/main/java/dev/vality/fistful/magista/util/TokenStringUtil.java
@@ -28,9 +28,9 @@ public class TokenStringUtil {
             if (params.length != 3) {
                 throw new IllegalArgumentException("Invalid token");
             }
-            long queryHashCode = Integer.parseInt(params[0]);
-            long idHashCode = Integer.parseInt(params[1]);
-            if ((long) queryParameters.hashCode() != queryHashCode || params[2].hashCode() != idHashCode) {
+            int queryHashCode = Integer.parseInt(params[0]);
+            int idHashCode = Integer.parseInt(params[1]);
+            if (queryParameters.hashCode() != queryHashCode || params[2].hashCode() != idHashCode) {
                 throw new IllegalArgumentException("Invalid token");
             }
         }
